text: xml: add tests for if/unless attributes
