+++ CMakeLists.txt
@@ -200,6 +200,7 @@ if(CATKIN_ENABLE_TESTING)
 
 		catch_add_test(test_xml_loading
 			test/xml/test_param.cpp
+			test/xml/test_if_unless.cpp
 		)
 		target_link_libraries(test_xml_loading
 			rosmon_launch_config
