+++ test/xml/test_param.cpp
@@ -1,65 +1,13 @@
+// Unit tests for <param> tags
+// Author: Max Schwarz <max.schwarz@ais.uni-bonn.de>
 
 #include <catch_ros/catch.hpp>
 
 #include "../../src/launch/launch_config.h"
 
+#include "param_utils.h"
+
 using namespace rosmon::launch;
-
-typedef std::map<std::string, XmlRpc::XmlRpcValue> ParameterMap;
-
-namespace Catch
-{
-	template<>
-	struct StringMaker<XmlRpc::XmlRpcValue>
-	{
-		static std::string convert(const XmlRpc::XmlRpcValue& value)
-		{
-			// We need a copy, since the cast operators don't work on const
-			// refs *sigh*
-			XmlRpc::XmlRpcValue copy = value;
-
-			std::stringstream ss;
-			ss << "XmlRpc";
-			switch(value.getType())
-			{
-				case XmlRpc::XmlRpcValue::TypeInt:
-					ss << "<int>(" << static_cast<int>(copy) << ")";
-					break;
-				case XmlRpc::XmlRpcValue::TypeBoolean:
-					ss << "<bool>(" << static_cast<bool>(copy) << ")";
-					break;
-				case XmlRpc::XmlRpcValue::TypeString:
-					ss << "<string>('" << static_cast<std::string>(copy) << "')";
-					break;
-				case XmlRpc::XmlRpcValue::TypeDouble:
-					ss << "<double>(" << static_cast<double>(copy) << ")";
-					break;
-				default:
-					ss << "<unknown>()";
-					break;
-			}
-
-			return ss.str();
-		}
-	};
-
-	template<>
-	struct StringMaker<ParameterMap>
-	{
-		static std::string convert(const ParameterMap& value)
-		{
-			std::stringstream ss;
-			ss << "{";
-			for(auto& param : value)
-			{
-				ss << "\"" << param.first << "\"=" << StringMaker<XmlRpc::XmlRpcValue>::convert(param.second) << ", ";
-			}
-			ss << "}";
-
-			return ss.str();
-		}
-	};
-}
 
 TEST_CASE("global_param", "[param]")
 {
@@ -80,21 +28,6 @@ TEST_CASE("global_param", "[param]")
 
 	XmlRpc::XmlRpcValue value = it->second;
 	REQUIRE(static_cast<std::string>(value) == "hello_world");
-}
-
-template<class T>
-void checkTypedParam(const ParameterMap& parameters, const std::string& name, XmlRpc::XmlRpcValue::Type expectedType, T expected)
-{
-	CAPTURE(parameters);
-	CAPTURE(name);
-
-	auto it = parameters.find(name);
-	REQUIRE(it != parameters.end());
-
-	XmlRpc::XmlRpcValue value = it->second;
-
-	REQUIRE(value.getType() == expectedType);
-	REQUIRE(static_cast<T>(value) == expected);
 }
 
 TEST_CASE("param_types", "[param]")
